Removed unneeded functions
